model_requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
albumentations>=1.0.0
tqdm>=4.62.0
wandb>=0.12.0
scipy>=1.7.0
pandas>=1.3.0